# backend/src/main/java/com/anveshak/DTOs/UpdatePaperRequest.java
package com.anveshak.DTOs;

public record UpdatePaperRequest(
        String title,
        String abstractText,
        String[] authors,
        Integer publicationYear,
        String[] keywords) {

}