# backend/src/main/resources/openapi.yml
openapi: 3.1.0
info:
  title: Anveshak API
  version: 1.0.0
  description: API documentation for the Anveshak research paper management application.
servers:
  - url: /
...
security:
  - bearerAuth: []
paths:
...
  /papers/{paperId}:
    get:
      tags: [Research Paper Management]
      operationId: getPaper
      summary: Get a research paper by ID
      parameters:
        - name: paperId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Paper details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResearchPaperResponse'
    patch:
      tags: [Research Paper Management]
      operationId: updatePaper
      summary: Update a research paper and optionally replace its PDF
      parameters:
        - name: paperId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [paper]
              properties:
                paper:
                  $ref: '#/components/schemas/UpdatePaperRequest'
                pdfFile:
                  type: string
                  format: binary
            encoding:
              paper:
                contentType: application/json
      responses:
        '200':
          description: Paper updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResearchPaperResponse'
    delete:
      tags: [Research Paper Management]
      operationId: deletePaper
      summary: Delete a research paper
      parameters:
        - name: paperId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Paper deleted
...
  /collections:
    post:
      tags: [Collection Management]
      operationId: createCollection
      summary: Create a collection
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NewCollectionRequest'
      responses:
        '201':
          description: Collection created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResearchCollectionResponse'
    get:
      tags: [Collection Management]
      operationId: listCollections
      summary: List collections
      responses:
        '200':
          description: Collections
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ResearchCollectionResponse'
  /collections/{collectionId}:
    get:
      tags: [Collection Management]
      operationId: getCollection
      summary: Get a collection by ID
      parameters:
        - name: collectionId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Collection details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResearchCollectionResponse'
    patch:
      tags: [Collection Management]
      operationId: updateCollection
      summary: Update a collection
      parameters:
        - name: collectionId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateCollectionRequest'
      responses:
        '200':
          description: Collection updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResearchCollectionResponse'
    delete:
      tags: [Collection Management]
      operationId: deleteCollection
      summary: Delete a collection
      parameters:
        - name: collectionId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Collection deleted
  /collections/{collectionId}/papers/{paperId}:
    post:
      tags: [Collection Management]
      operationId: addPaperToCollection
      summary: Add a paper to a collection
      parameters:
        - name: collectionId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: paperId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Collection updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResearchCollectionResponse'
    delete:
      tags: [Collection Management]
      operationId: removePaperFromCollection
      summary: Remove a paper from a collection
      parameters:
        - name: collectionId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: paperId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Collection updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResearchCollectionResponse'
  /collections/{collectionId}/papers:
    get:
      tags: [Collection Management]
      operationId: listCollectionPapers
      summary: List papers in a collection
      parameters:
        - name: collectionId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Papers in the collection
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ResearchPaperResponse'
...
  /chat-sessions:
    post:
      tags: [Paper Chat]
      operationId: createChatSession
      summary: Create a chat session for a paper
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NewChatSessionRequest'
      responses:
        '201':
          description: Chat session created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ChatSessionResponse'
    get:
      tags: [Paper Chat]
      operationId: listChatSessions
      summary: List chat sessions for the current user
      responses:
        '200':
          description: Chat sessions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ChatSessionResponse'
  /chat-sessions/{sessionId}:
    delete:
      tags: [Paper Chat]
      operationId: deleteChatSession
      summary: Delete a chat session
      parameters:
        - name: sessionId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Chat session deleted
  /chat-sessions/{sessionId}/messages:
    get:
      tags: [Paper Chat]
      operationId: listChatMessages
      summary: List messages in a chat session
      parameters:
        - name: sessionId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Chat messages
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ChatMessageResponse'
    post:
      tags: [Paper Chat]
      operationId: sendChatMessage
      summary: Send a message to a chat session
      parameters:
        - name: sessionId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ChatMessageRequest'
      responses:
        '200':
          description: Assistant response message
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ChatMessageResponse'
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: jwt
  schemas:
...
    NewCollectionRequest:
      type: object
      required: [name]
      properties:
        name:
          type: string
    UpdateCollectionRequest:
      type: object
      properties:
        name:
          type: string
    ResearchCollectionResponse:
      type: object
      required: [id, name, createdAt, papers]
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        createdAt:
          type: string
          format: date-time
        papers:
          type: array
          items:
            $ref: '#/components/schemas/ResearchPaperResponse'
...
    UpdatePaperRequest:
      type: object
      properties:
        title:
          type: string
        abstractText:
          type: string
        authors:
          type: array
          items:
            type: string
        publicationYear:
          type: integer
        keywords:
          type: array
          items:
            type: string
    ResearchPaperResponse:
      type: object
      required: [id, title, abstractText, authors, keywords, createdAt, updatedAt]
      properties:
        id:
          type: string
          format: uuid
        title:
          type: string
        abstractText:
          type: string
        authors:
          type: array
          items:
            type: string
        keywords:
          type: array
          items:
            type: string
        publicationYear:
          type: integer
          nullable: true
        pdfUrl:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
    NewChatSessionRequest:
      type: object
      required: [paperId]
      properties:
        paperId:
          type: string
    ChatSessionResponse:
      type: object
      required: [sessionId]
      properties:
        sessionId:
          type: string
    ChatMessageRequest:
      type: object
      required: [message, sessionId, role]
      properties:
        message:
          type: string
        sessionId:
          type: string
        role:
          type: string
          description: Message role provided by the caller.
    ChatMessageResponse:
      type: object
      required: [message, sessionId]
      properties:
        message:
          type: string
        sessionId:
          type: string
...